Update mconvert and vision scripts

+++ scripts/vision-radio-dl.py
@@ -9,7 +9,7 @@ from xml.etree import ElementTree
 
 RSS_LINK = "https://feeds.castos.com/9n6dg"
 
-logging.basicConfig(level=logging.INFO)
+logging.basicConfig(level=logging.INFO, format="%(levelname)s %(asctime)s %(message)s")
 logger = logging.getLogger(__name__)
 
 parser = argparse.ArgumentParser(
@@ -24,6 +24,10 @@ parser.add_argument(
 parser.add_argument(
     "-l", "--limit", nargs="?", type=int, default=5,
     help="Podcasts count to download",
+)
+parser.add_argument(
+    "-o", "--overwrite", type=bool, default=False,
+    help="Overwrite existing files",
 )
 
 
@@ -40,13 +44,19 @@ def prepare_data() -> list[dict]:
     ]
 
 
-def download(target: str, limit: int):
+def download(target: str, limit: int, overwrite: bool = False):
     podcasts = prepare_data()
     for podcast in sorted(podcasts, key=lambda x: x["pub_date"], reverse=True)[:limit]:
+        target_file_name = f"{target}/{podcast['filename']}.mp3"
+        if os.path.exists(target_file_name) and not overwrite:
+            logger.warning(f"Podcast {podcast['filename']} already exists")
+            continue
+
         response = requests.get(podcast["link"])
         if not response.ok:
             logger.error(f"Failed to download a file {podcast['filename']} - {response.reason}")
             continue
+
         with open(f"{target}/{podcast['filename']}.mp3", "wb") as f:
             f.write(response.content)
             logger.info(f"Podcast {podcast['filename']} successfully downloaded")
@@ -54,4 +64,4 @@ def download(target: str, limit: int):
 
 if __name__ == "__main__":
     args = parser.parse_args()
-    download(args.target, args.limit)
+    download(args.target, args.limit, args.overwrite)
